Moving the thresholding functionality into the utilities directory.
